fix(anon,band3,system): lane CC-4 — 18 anon names + 8 source fixes, +14 honest
Coordinator A/B in an own worktree off d459ce15, config.yml touched on BOTH legs
(a map edit is a silent no-op without a re-split):

                default                        name_check
  matched  41,387 -> 41,401  (+14)      41,385 -> 41,399
  honest   39,899 -> 39,913  (+14)      39,897 -> 39,911
  code%    35.701050 -> 35.725967       23.099808 -> 23.101019
  matched_code 3,777,184 -> 3,779,820 B (+2,636 B)

  strict set: LOST 0 / GAINED 14 -- zero regressions, not a net figure.
  masked_equal flat at 1,488 => supply-backed, not oversubscription.
  total_functions 69,367 on every leg (same-split).
  duplicate-key gate: 27,581 pairs / 27,581 unique, PASS.

The +2,636 B equals EXACTLY the sum of the eight functions taken byte-identical:
SongParser::Reset 960, EndingBonus::SyncObjects 536, BandMachineMgr::
RemoveRemoteMachine 316, Gem::operator= 244, XboxEnumeration::ctor 200,
BandMachineMgr::GetMachines 156, QuestFilterProvider::Update 128,
Track::UserName 96.

THE VEIN WAS 9x SMALLER THAN THE BRIEF, and sizing it is the more durable result.
Briefed as 1,084 functions at >=95% similarity reporting 0.0; measured as 119 rows
/ 17,792 B. The anon-0.0 population is 5,656 supplied, not 6,260 -- the difference
is the ~504 names BV-4/BW-1/BW-2 already landed. ?Handle@Rnd@@ was already taken
by BW-2. Disposition of all 119: 29 uncontested+unambiguous, 7 dropped on name
collisions, 22 adjudicated, 18 ACCEPT / 3 REJECT / 1 UNDECIDED. The 90
contested/ambiguous rows swept at agree>=2 shipped exactly zero. Top penalty in
the whole pool was 10.62 instruction-words. The fuzzy >=95% vein is DRAINED.

Naming the 18 rows paid +6 honest and +0.000000 pp code%, confirming to the digit
that naming cannot move code% (matched_code needs raw fuzzy == 100). That yields a
reusable discriminator, which held across all 18:

    norm >  fuzzy  =>  regalloc-only  -- pays on naming alone
    norm == fuzzy  =>  real source divergence -- needs a fix

A PRE-FLIP VERDICT WAS WRONG, and this is the part worth carrying forward.
BandMachineMgr::RemoveRemoteMachine was filed as an unfixable srawi./clrrwi.
build_env wall. That verdict predates the cl-10224 compiler flip, and the real
cause was a porting artifact: a stray (int) cast absent from the rb3-Wii oracle.
Other at_limit/build_env rows carrying that signature are worth re-triaging --
but NOT as a blanket sweep: 3 of the 5 casts in that same file are load-bearing
at 100%.

Instrument honesty, recorded rather than glossed: anon_reloc_cmp.py structurally
could not adjudicate this pool (it hard-returns NOT_BYTE_PERFECT before evaluating
evidence). After separating the gate -- byte-perfection tests the BODY,
adjudicate() tests the REFERENTS -- truth-ablation gave POS 6/6, NEG 0/29. That is
underpowered: a 95% lower bound on precision of 60.7%, versus 99.92% for the
byte-perfect population, and the plant rate is bounded at <=9.8%, not zero. 14 of
18 names are corroborated by a second instrument (reached norm-100, 8 of them
byte-identically) and a wrong name cannot be repaired to byte-identity by a
one-line change; 4 rest on the reloc gate alone.

Refused: 46 rows passing the full reloc gate, all byte-perfect -- the already-
drained class, whose population this lane's control does not calibrate, and 20+ of
which are the ?NewObject@X@@ family where the StaticClassName channel previously
found 24 false-100s. Withheld but NOT condemned: 7 name collisions, including a
push_back trio spanning three different units, which are plausibly legitimate
per-unit instantiations refused only because the applier asserts globally.

+++ src/band3/bandtrack/Gem.h
@@ -68,7 +68,7 @@ public:
     float mArrhythmicDurationSeconds; // 0x4c
     bool mInArrhythmic; // 0x50
     int unk_0x58; // 0x54 (copied as a word by operator=)
-    int unk_0x40; // 0x58
+    float unk_0x40; // 0x58 — retail copies this with lfs/stfs in operator=, so it is a float
     bool mIsCymbalLane; // 0x5c — gem is on a game-cymbal lane
     unsigned int unk_0x44; // 0x60 - some RG chord shape
     unsigned int unk_0x48; // 0x64 - some other RG chord shape
